Refactor SidebarNavigationRoot to use NavigatorScreen
