+++ packages/edit-site/src/components/sidebar-navigation-root/index.js
@@ -1,86 +1,17 @@
 /**
  * WordPress dependencies
  */
-import {
-	__experimentalItemGroup as ItemGroup,
-	__experimentalVStack as VStack,
-	__experimentalHStack as HStack,
-	Button,
-} from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-import { layout, symbolFilled } from '@wordpress/icons';
-import { useDispatch } from '@wordpress/data';
-import { useViewportMatch } from '@wordpress/compose';
+import { __experimentalNavigatorProvider as NavigatorProvider } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
-import SidebarNavigationTitle from '../sidebar-navigation-title';
-import { useLink } from '../routes/link';
-import SidebarNavigationItem from '../sidebar-navigation-item';
-import { useLocation } from '../routes';
-import { store as editSiteStore } from '../../store';
-import getIsListPage from '../../utils/get-is-list-page';
+import SidebarNavigationScreenMain from './screen-main';
 
 export default function SidebarNavigationRoot() {
-	const { params } = useLocation();
-	const isListPage = getIsListPage( params );
-	const isEditorPage = ! isListPage;
-	const templatesLink = useLink( {
-		postType: 'wp_template',
-		postId: undefined,
-	} );
-	const templatePartsLink = useLink( {
-		postType: 'wp_template_part',
-		postId: undefined,
-	} );
-	const { __unstableSetCanvasMode } = useDispatch( editSiteStore );
-	const isMobileViewport = useViewportMatch( 'medium', '<' );
-	const isTemplatesPage =
-		params.postType === 'wp_template' && ! params.postId;
-	const isTemplatePartsPage =
-		params.postType === 'wp_template_part' && ! params.postId;
-
 	return (
-		<VStack spacing={ 6 }>
-			<SidebarNavigationTitle
-				title={
-					<HStack style={ { minHeight: 36 } }>
-						<div>{ __( 'Design' ) }</div>
-						{ ! isMobileViewport && isEditorPage && (
-							<Button
-								className="edit-site-layout__edit-button"
-								label={ __( 'Open the editor' ) }
-								onClick={ () => {
-									__unstableSetCanvasMode( 'edit' );
-								} }
-							>
-								{ __( 'Edit' ) }
-							</Button>
-						) }
-					</HStack>
-				}
-			/>
-			<nav className="edit-site-sidebar-navigation-root">
-				<ItemGroup>
-					<SidebarNavigationItem
-						{ ...templatesLink }
-						icon={ layout }
-						aria-current={ isTemplatesPage ? 'page' : undefined }
-					>
-						{ __( 'Templates' ) }
-					</SidebarNavigationItem>
-					<SidebarNavigationItem
-						{ ...templatePartsLink }
-						icon={ symbolFilled }
-						aria-current={
-							isTemplatePartsPage ? 'page' : undefined
-						}
-					>
-						{ __( 'Template Parts' ) }
-					</SidebarNavigationItem>
-				</ItemGroup>
-			</nav>
-		</VStack>
+		<NavigatorProvider initialPath="/">
+			<SidebarNavigationScreenMain />
+		</NavigatorProvider>
 	);
 }
